requirements.txt
blpapi>=3.24.0
dash>=2.14.0
plotly>=5.18.0
numpy>=1.26.0
scipy>=1.11.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
simple-websocket>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
